Changed to Ternary Operators.
For performance and code readability I have changed all the conditions checks to ternary operators. 
Looking for way to improve performance by performing the conditional check on the amount and actually having amount
be equal to that. A check and exception in the same line. For now this is a reusable piece of software.